+++ src/main/java/com/schedule/calendar/Controllers/UserController.java
@@ -1,10 +1,12 @@
 package com.schedule.calendar.Controllers;
 
 import com.schedule.calendar.Models.User;
+import com.schedule.calendar.Models.UserRole;
+import com.schedule.calendar.Models.UserType;
 import com.schedule.calendar.Repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCrypt;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -20,8 +22,7 @@ public class UserController {
     @Autowired
     private UserRepository userRepository;
     @Autowired
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
-
+    private PasswordEncoder passwordEncoder;
 
     @GetMapping("/signup")
     public ModelAndView addUser() {
@@ -32,12 +33,15 @@ public class UserController {
     }
 
     @PostMapping("/signup")
-    public String saveUser(@RequestBody User user) {
+    @ResponseStatus(HttpStatus.CREATED)
+    public User saveUser(@RequestBody User user) {
 //        String encodedPassword = BCrypt.hashpw(user.getPassword(), BCrypt.gensalt(12));
 //        user.setPassword(encodedPassword);
-        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
-        this.userRepository.save(user);
-        return "calendar";
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        user.setUserRole(UserRole.USER);
+        user.setUserType(UserType.DEFAULT);
+        return  userRepository.save(user);
+       
     }
 
     @GetMapping("/login")
