--- src/main/java/com/schedule/calendar/Models/MyUserDetailsService.java
@@ -1,43 +1,0 @@
-package com.schedule.calendar.Models;
-
-import com.schedule.calendar.Repositories.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class MyUserDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-    public MyUserDetailsService(UserRepository userRepository) {
-    this.userRepository = userRepository;
-    }
-    
-    /**
-     * Locates the user based on the username. In the actual implementation, the search
-     * may possibly be case sensitive, or case insensitive depending on how the
-     * implementation instance is configured. In this case, the <code>UserDetails</code>
-     * object that comes back may have a username that is of a different case than what
-     * was actually requested..
-     *
-     * @param username the username identifying the user whose data is required.
-     * @return a fully populated user record (never <code>null</code>)
-     * @throws UsernameNotFoundException if the user could not be found or the user has no
-     *                                   GrantedAuthority
-     */
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-      Optional<User> user = userRepository.findByUsername(username);
-      if (user.isPresent()) {
-          var userObj = user.get();
-          return org.springframework.security.core.userdetails.User.builder()
-                  .username(userObj.getUsername())
-                  .password(userObj.getPassword())
-                  .roles(String.valueOf(userObj.getRole()))
-                  .build();
-      }
-      else throw new UsernameNotFoundException(username);
-    }
-}
