+++ src/main/java/com/schedule/calendar/Models/UserRole.java
@@ -1,9 +1,6 @@
 package com.schedule.calendar.Models;
 
 public enum UserRole {
-    STUDENT,
-    EMPLOYEE,
-    ORGANIZATION,
     USER,
     ADMIN
 }
