+++ src/main/java/com/schedule/calendar/Models/UserType.java
@@ -1,0 +1,8 @@
+package com.schedule.calendar.Models;
+
+public enum UserType {
+    DEFAULT,
+    STUDENT,
+    EMPLOYEE,
+    ORGANIZATION,
+}
