+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.datasource.url=jdbc:mysql://localhost:3306/calendar?createDatabaseIfNotEx
 spring.datasource.username=root
 spring.datasource.password=secret
 spring.jpa.hibernate.ddl-auto=update
+spring.jpa.open-in-view=true
